+++ frontend/src/components/cards/OpenFoodFactsProductCard.css
